% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krlift-package.R
\docType{package}
\name{krlift-package}
\alias{krlift}
\alias{krlift-package}
\title{krlift: Lift Status in Korea}
\description{
Lift Status in Korea from https://www.data.go.kr.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/mrchypark/krlift}
  \item Report bugs at \url{https://github.com/mrchypark/krlift/issues}
}

}
\author{
\strong{Maintainer}: ChanYub Park \email{mrchypark@gmail.com}

}
\keyword{internal}
